requirements.txt
pytest
kafka-python==2.0.2
schedule==1.1.0
Apache-TrafficControl==3.0.0
requests==2.27.1
PyYaml==6.0
elastic-apm==6.8.1
git+https://github.com/ARMmaster17/watergrid-python.git@main